scdds1065 - Changed the extent construction used in the database.
Previously, the extent was a construct that acted as a proxy to the type. This approach had several drawbacks, but mainly caused misuse of the type (keeps on the type (logical) while the extent should actually be kept). This is now fixed as follows:

The extent is now a subclass of a typedef of type, allowing for transparant usage of extents and multiple extents for one type. Eventually both the c_new() and c_free() should transparantly support extents. Now only c_free() recognizes memory allocated through an extent, freeing it through the extents as expected. Allocation through an extent now still has to be carried out explicitly by calling c_extentCreate().

  |------|
  | type |
  |------|
     ^
     |
 |---------|
 | typedef |
 |---------|
     ^
     |
 |--------|
 | extent |
 |--------|

The baseInit() bootstrap has been simplified by this change and refcounting now works on types (and extents, cause an extent is a type) as expected.