scdds456 - Implemented statistics on blocked/timedout writes.
The numberOfBlockedWrites statistic that used to exist (but was not implemented)
has been implemented by means of the following statistics on the writer:

- numberOfTimedOutWrites:
    Counts all writes that return with a timeout.
- numberOfWritesBlockedBySamplesLimit:
    Counts all writes that are blocked (at least once) on the max_samples re-
    source limit. The write may or may not have returned with a timeout.
- numberOfWritesBlockedByInstanceLimit:
    Counts all writes that are blocked (at least once) on the max_instances re-
    source limit. The write may or may not have returned with a timeout.
- numberOfWritesBlockedBySamplesPerInstanceLimit:
    Counts all writes that are blocked (at least once) on the
    max_samples_per_instance resource limit. The write may or may not have
    returned with a timeout.

So when a write succeeds, but got blocked during the process on a resource limit,
the respective counter(s) will be incremented. When a write got blocked and even-
tually times out, the respective counter(s) and the numberOfTimedOutWrites will
be incremented.