scdds1433 - Implemented an extension to the memory-manager for caches.
When a cache is malloc'd through the new c_mmMallocCache and freed
through c_mmFreeCache, the memory manager will maintain a list of
all caches, which can be traversed when memory usage of caches is
wanted. Performance-wise, a malloc for a cache got constant-time
more expensive, a free can be more expensive (as the cache has to
be found in the list).