scdds734 - Added a counting mechanism to the user-layer and modified detaching.
Added an init-counting mechanism to support multiple init- and detaching
and changed the way the userlayer was detached. During a detach, the
user-lock is not released anymore, ensuring that no process can be
working on a partially deinitialized user-admin object.

Also moved os-layer initialization code specific for Windows to the OS-layer,
which is made possible by the counting mechanism added to the os-init/exit.