fixes issue 131
Built with MSVC 14.2 (only).
Added copyIn and copyOut declarations for typedefs to match the declarations used for structs.
Added for both genSpliceType and genISOCxx2SpliceType, not sure if there are other places to add it.
Removed unnecessary references to structs and classes on the param types from the struct code, because the typedefs don't just deal with structs, but I'm not sure if this will cause issues with GCC etc.